pyproject.toml
[project]
name = "junos-exporter"
version = "0.1.0"
description = "Junos Exporter using PyEZ Tables and Views"
authors = [
  {name = "minefuto", email = "minefuto@gmail.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "junos-eznc>=2.7.2",
    "ntc-templates>=7.7.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Repository = "https://github.com/minefuto/junos-exporter.git"

[project.scripts]
junos-exporter = "junos_exporter.cli:cli"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.9.7",
    "types-pyyaml>=6.0.12.20241230",
]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff.lint.pyflakes]
allowed-unused-imports = [
    "jnpr.junos.op.arp",
    "jnpr.junos.op.bfd",
    "jnpr.junos.op.bgp",
    "jnpr.junos.op.ccc",
    "jnpr.junos.op.ddos",
    "jnpr.junos.op.elsethernetswitchingtable",
    "jnpr.junos.op.ethernetswitchingtable",
    "jnpr.junos.op.ethport",
    "jnpr.junos.op.fpc",
    "jnpr.junos.op.idpattacks",
    "jnpr.junos.op.intopticdiag",
    "jnpr.junos.op.inventory",
    "jnpr.junos.op.isis",
    "jnpr.junos.op.l2circuit",
    "jnpr.junos.op.lacp",
    "jnpr.junos.op.ldp",
    "jnpr.junos.op.lldp",
    "jnpr.junos.op.nd",
    "jnpr.junos.op.ospf",
    "jnpr.junos.op.pfestats",
    "jnpr.junos.op.phyport",
    "jnpr.junos.op.ppm",
    "jnpr.junos.op.routes",
    "jnpr.junos.op.securityzone",
    "jnpr.junos.op.systemstorage",
    "jnpr.junos.op.taskmemory",
    "jnpr.junos.op.teddb",
    "jnpr.junos.op.vlan",
    "jnpr.junos.op.xcvr",
]
